Fix counter at the top of Summary tab (#193)
